+++ extensions/advanced-replace/src/utilities/storage.ts
@@ -17,6 +17,21 @@ export async function setSavedItems(options: Entry[] | undefined) {
   return LocalStorage.setItem("regexOptions", JSON.stringify(options));
 }
 
+export async function moveItem(fromIndex: number, toIndex: number, callbackFn?: () => void) {
+  const savedItems = await getSavedItems();
+  const len = savedItems.length;
+  if (len === 0 || fromIndex === toIndex) return;
+  if (fromIndex < 0 || fromIndex >= len || toIndex < 0 || toIndex >= len) return;
+
+  const updatedEntries = savedItems.slice();
+  const [movedItem] = updatedEntries.splice(fromIndex, 1);
+  if (movedItem === undefined) return;
+  updatedEntries.splice(toIndex, 0, movedItem);
+
+  await setSavedItems(updatedEntries);
+  callbackFn?.();
+}
+
 export async function deleteSavedItem(item: Entry) {
   if (
     await confirmAlert({